% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_conditions}
\alias{aa_conditions}
\title{List of alcohol attributable conditions}
\format{data frame with 69 rows and 8 fields

\preformatted{
'data.frame':	69 obs. of  8 variables:
$ cat1          : Factor w/ 3 levels "Partially attributable conditions - acute conditions",..: 3 3 3 3 3 3 3 3 3 3 ...
$ cat2          : Factor w/ 12 levels "Cardiovascular disease",..: 12 12 12 12 12 12 12 12 12 12 ...
$ desc          : Factor w/ 68 levels "§","§§","Accidental excessive cold",..: 9 49 21 14 13 10 11 12 7 8 ...
$ attribution   : logi  NA NA NA NA NA NA ...
$ cause         : logi  NA NA NA NA NA NA ...
$ codes         : Factor w/ 68 levels "A15-A19","C00-C14",..: 11 12 13 15 16 19 27 28 34 35 ...
$ condition_fuid: Ord.factor w/ 69 levels "Wholly attributable conditions.Wholly attributable conditions.Alcohol-induced pseudo-Cushing's syndrome.E24.4"<..: 1 2 3 4 5 6 7 8 9 10 ...
$ condition_uid : int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
aa_conditions
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
require("dplyr")
aa_conditions \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_"), -codes, -desc) \%>\%
    summary(16)

}
\keyword{datasets}
