% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_versions}
\alias{aa_versions}
\title{List of alcohol attributable conditions by version}
\format{data frame with 185 rows and 2 fields

\preformatted{
Class 'data.frame':	185 obs. of  2 variables:
$ version      : Factor w/ 4 levels "aaf_2007_ni39",..: 1 1 1 1 1 1 1 1 1 1 ...
$ condition_uid: int  1 2 3 4 5 6 7 10 13 14 ...
}}
\usage{
aa_versions
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
if (isNamespaceLoaded("dplyr")) {
  require("dplyr")

  aafractions.ncc::aa_versions \%>\%
    dplyr::mutate_if(is.character, as.factor) \%>\%
    dplyr::select(-starts_with("condition_")) \%>\%
    summary(16)
}

}
\seealso{
Other alcohol datasets: \code{\link{aa_conditions}},
  \code{\link{aa_fractions}},
  \code{\link{alcohol_datasets}},
  \code{\link{lu_aac_icd10}}, \code{\link{sa_versions}}
}
\concept{alcohol datasets}
\keyword{datasets}
