% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aafractions.R
\docType{package}
\name{aafractions.ncc}
\alias{aafractions.ncc}
\alias{aafractions.ncc-package}
\title{aafractions.ncc: Alcohol Attributable Fractions Lookup Tables}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\section{Foo functions}{

The foo functions ...
}

