% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{ab_labels_from_breaks}
\alias{ab_labels_from_breaks}
\title{Create ageband labels}
\usage{
ab_labels_from_breaks(breaks)
}
\arguments{
\item{breaks}{(integer vector) breaks}
}
\value{
(character vector) labels
}
\description{
Given breaks construct ageband labels of the form "xx-yy Yrs" and "zz+
Yrs"
}
\seealso{
Other examples_of_analysis: \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}}, \code{\link{gen_codes}},
  \code{\link{main__example_analysis__hes_ip}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
