% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ac_attribution}
\alias{ac_attribution}
\title{List of Ambulatory care sensitive fractions by condition and age}
\format{data frame with 13 rows and 3 fields

\preformatted{
Observations: 13
Variables: 3
$ condition_uid <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13
$ version       <chr> "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_2...
$ acs_af        <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
}}
\usage{
ac_attribution
}
\description{
Provides lookup tables for use with Ambulatory care sensitive analyses.
}
\examples{
if (all(sapply(c("dplyr", "reshape2"), isNamespaceLoaded))) {
  require("dplyr", warn.conflicts = FALSE)
  require("reshape2", warn.conflicts = FALSE)

  ac_attribution \%>\%
    dcast(
      condition_uid + condition_description ~ ab_ucs_explode
      , value.var = "ucs_af", fill = 0
    )
}


}
\seealso{
Other ambulatory care sensitive datasets: \code{\link{ac_conditions}},
  \code{\link{ac_versions}}, \code{\link{acs_datasets}},
  \code{\link{lu_acc_icd10_sec}},
  \code{\link{lu_acc_icd10}}, \code{\link{lu_acc_opcs}}
}
\concept{ambulatory care sensitive datasets}
\keyword{datasets}
