% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ac_versions}
\alias{ac_versions}
\title{List of Ambulatory care sensitive conditions by version}
\format{data frame with 13 rows and 2 fields

\preformatted{
Observations: 13
Variables: 2
$ version       <chr> "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_26", "ccg_ois_2...
$ condition_uid <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13
}}
\usage{
ac_versions
}
\description{
Provides lookup tables for use with Ambulatory care sensitive analyses.
}
\examples{
if (isNamespaceLoaded("dplyr")) {
  require("dplyr")

  aafractions.ncc::ac_versions \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_")) \%>\%
    summary(16)
}


}
\seealso{
Other ambulatory care sensitive datasets: \code{\link{ac_attribution}},
  \code{\link{ac_conditions}}, \code{\link{acs_datasets}},
  \code{\link{lu_acc_icd10_sec}},
  \code{\link{lu_acc_icd10}}, \code{\link{lu_acc_opcs}}
}
\concept{ambulatory care sensitive datasets}
\keyword{datasets}
