% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{acs_datasets}
\alias{acs_datasets}
\title{Ambulatory care sensitive Emergency datasets}
\description{
Ambulatory care sensitive Emergency datasets
}
\seealso{
Other ambulatory care sensitive datasets: \code{\link{ac_attribution}},
  \code{\link{ac_conditions}}, \code{\link{ac_versions}},
  \code{\link{lu_acc_icd10}}
}
\concept{ambulatory care sensitive datasets}
