% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{alcohol_datasets}
\alias{alcohol_datasets}
\title{Alcohol datasets}
\description{
Alcohol datasets
}
\seealso{
Other alcohol datasets: \code{\link{aa_conditions}},
  \code{\link{aa_fractions}}, \code{\link{aa_versions}},
  \code{\link{lu_aac_icd10}}
}
\concept{alcohol datasets}
