% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_hesip.R
\name{create__dummy_hesip}
\alias{create__dummy_hesip}
\title{Create dummy HES table}
\usage{
create__dummy_hesip(n = 1024, bWriteCSV = FALSE)
}
\arguments{
\item{n}{(integer) length of table}

\item{bWriteCSV}{(bool) save to disk .. or not}
}
\description{
Generate a psudo random HES IP table good enough to put any analysis through
its paces.
}
