% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_hesip.R
\name{create_lu_ageband}
\alias{create_lu_ageband}
\title{Create age band looks for aa and esp}
\usage{
create_lu_ageband()
}
\value{
(data.frame) with fields
}
\description{
Create age band looks for aa and esp
}
\examples{
require("dplyr")
aafractions.ncc:::create_lu_ageband() \%>\% mutate_if(is.character, as.factor) \%>\% summary(20)
aafractions.ncc:::create_lu_ageband() \%>\% select(starts_with("ab_")) \%>\% unique()

}
