% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{examples_of_analysis}
\alias{examples_of_analysis}
\title{Example analysis}
\description{
Example analysis
}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}}, \code{\link{gen_codes}},
  \code{\link{main__example_analysis__hes_ip}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
