% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{ilike}
\alias{ilike}
\title{Case insensitive like}
\usage{
ilike(vector, pattern)
}
\arguments{
\item{vector}{(character) string to match}

\item{pattern}{(character) passed to \code{grepl}}
}
\value{
(logical vector) indicate matches
}
\description{
Case insensitive like
}
\details{
Inspiration from data.table \code{like}.
}
