% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_aac_icd10}
\alias{lu_aac_icd10}
\title{Lookup between alcohol attributable conditions and icd10 codes}
\format{data frame with 69 rows and 7 fields}
\usage{
lu_aac_icd10
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
require("dplyr")

merge(lu_aac_icd10, aa_versions, by = "condition_uid") \%>\%
  dcast(
    ... ~ Version, value.var = "condition_uid", fun = paste, collapse = "|"
  ) \%>\%
  arrange(icd10)

}
\keyword{datasets}
