% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_acc_icd10}
\alias{lu_acc_icd10}
\title{Lookup between Ambulatory care sensitive conditions and icd10 codes}
\format{data frame with 277 rows and 2 fields

\preformatted{
Observations: 277
Variables: 2
$ condition_uid <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3,...
$ icd10         <chr> "I480", "I481", "I482", "I483", "I484", "I485", "I486", "I487", "I488", "I489", "I48X", "I48", "I200", "...
}}
\usage{
lu_acc_icd10
}
\description{
Provides lookup tables for use with Ambulatory care sensitive analyses.
Primary condition: primary diagnosis codes.
}
\examples{
if (all(sapply(c("dplyr", "reshape2"), isNamespaceLoaded))) {
  require("dplyr", warn.conflicts = FALSE)
  require("reshape2", warn.conflicts = FALSE)

  aafractions.ncc::lu_acc_icd10 \%>\%
    merge(aafractions.ncc::ac_versions, by = "condition_uid", allow.cartesian = TRUE) \%>\%
    dcast(... ~ version, value.var = "condition_uid", fun = paste, collapse = "|") \%>\%
    arrange(icd10) \%>\%
    head(16)
}

str(aafractions.ncc::lu_acc_icd10)

}
\seealso{
Other ambulatory care sensitive datasets: \code{\link{ac_attribution}},
  \code{\link{ac_conditions}}, \code{\link{ac_versions}},
  \code{\link{acs_datasets}}
}
\concept{ambulatory care sensitive datasets}
\keyword{datasets}
