% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_acc_icd10_sec}
\alias{lu_acc_icd10_sec}
\title{Lookup between Ambulatory care sensitive conditions and icd10 codes}
\format{data frame with x rows and x fields

\preformatted{
}}
\usage{
lu_acc_icd10_sec
}
\description{
Provides lookup tables for use with Ambulatory care sensitive analyses.
Secondary condition: secondary diagnosis codes.
}
\examples{

}
\seealso{
Other ambulatory care sensitive datasets: \code{\link{ac_attribution}},
  \code{\link{ac_conditions}}, \code{\link{ac_versions}},
  \code{\link{acs_datasets}}, \code{\link{lu_acc_icd10}},
  \code{\link{lu_acc_opcs}}
}
\concept{ambulatory care sensitive datasets}
\keyword{datasets}
