% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_fractions}
\alias{lu_fractions}
\title{List of alcohol attributable fractions by condition, age and sex}
\format{data frame with 5,920 rows and 7 fields

\preformatted{
'data.frame':	5920 obs. of  7 variables:
$ Version       : Factor w/ 4 levels "aaf_2007_ni39",..: 1 1 1 1 1 1 1 1 1 1 ...
$ condition_fuid: Ord.factor w/ 69 levels "Wholly attributable conditions.Wholly attributable conditions.Alcohol-induced pseudo-Cushing's syndrome.E24.4"<..: 1 1 1 1 1 1 1 1 1 1 ...
$ condition_uid : int  1 1 1 1 1 1 1 1 1 1 ...
$ aa_ageband    : chr  "00-15 Yrs" "16-24 Yrs" "25-34 Yrs" "35-44 Yrs" ...
$ sex           : chr  "F" "F" "F" "F" ...
$ analysis_type : chr  "morbidity" "morbidity" "morbidity" "morbidity" ...
$ aaf           : num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
lu_fractions
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
require("dplyr")
lu_fractions \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_")) \%>\%
    summary(16)

}
\keyword{datasets}
