% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_ucc_icd10}
\alias{lu_ucc_icd10}
\title{Lookup between Urgent care sensitive conditions and icd10 codes}
\format{data frame with 69 rows and 7 fields

\preformatted{
'data.frame':	5920 obs. of  6 variables:
$ version      : Factor w/ 4 levels "aaf_2007_ni39",..: 1 1 1 1 1 1 1 1 1 1 ...
$ condition_uid: int  1 1 1 1 1 1 1 1 1 1 ...
$ aa_ageband   : chr  "00-15 Yrs" "16-24 Yrs" "25-34 Yrs" "35-44 Yrs" ...
$ sex          : chr  "F" "F" "F" "F" ...
$ analysis_type: chr  "morbidity" "morbidity" "morbidity" "morbidity" ...
$ aaf          : num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
lu_ucc_icd10
}
\description{
Provides lookup tables for use with Urgent care sensitive analyses.
}
\examples{
if (all(sapply(c("dplyr", "reshape2"), isNamespaceLoaded))) {
  require("dplyr", warn.conflicts = FALSE)
  require("reshape2", warn.conflicts = FALSE)

  aafractions.ncc::lu_ucc_icd10 \%>\%
    merge(aafractions.ncc::uc_versions, by = "condition_uid", allow.cartesian = TRUE) \%>\%
    dcast(... ~ version, value.var = "condition_uid", fun = paste, collapse = "|") \%>\%
    arrange(icd10) \%>\%
    head(16)
}

}
\seealso{
Other urgent care sensitive datasets: \code{\link{uc_attribution}},
  \code{\link{uc_conditions}}, \code{\link{uc_versions}},
  \code{\link{ucs_datasets}}
}
\concept{urgent care sensitive datasets}
\keyword{datasets}
