% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_versions}
\alias{lu_versions}
\title{List of alcohol attributable fractions by condition, age and sex}
\format{data frame with 221 rows and 10 fields

\preformatted{
'data.frame':	213 obs. of  3 variables:
$ Version       : Factor w/ 4 levels "aaf_2007_ni39",..: 1 1 1 1 1 1 1 1 1 1 ...
$ condition_fuid: Ord.factor w/ 69 levels "Wholly attributable conditions.Wholly attributable conditions.Alcohol-induced pseudo-Cushing's syndrome.E24.4"<..: 1 2 3 4 5 6 7 10 13 14 ...
$ condition_uid : int  1 2 3 4 5 6 7 10 13 14 ...
}}
\usage{
lu_versions
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
lu_versions \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_") \%>\%
    summary(16)

}
\keyword{datasets}
