% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{main__example_analysis__aa_morbidity}
\alias{main__example_analysis__aa_morbidity}
\title{Example analysis}
\usage{
main__example_analysis__aa_morbidity()
}
\description{
Alcohol morbidity
}
\details{
- suitable for vignette
}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__sa_morbidity}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
