% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{main__example_analysis__uc_morbidity}
\alias{main__example_analysis__uc_morbidity}
\title{Example analysis}
\usage{
main__example_analysis__uc_morbidity(ip)
}
\arguments{
\item{ip}{(data frame) raw ip table to process}
}
\value{
(data frame) attributable events
}
\description{
Urgent care sensitive morbidity
}
\details{
- suitable for vignette

Methodology

Finished Admission Episodes (epiorder = 1, epistat = 3, patclass (1, 2, 5))
Emergency admission - admeth %like% '^2'
}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__aa_morbidity}},
  \code{\link{main__example_analysis__ac_morbidity}},
  \code{\link{main__example_analysis__sa_morbidity}},
  \code{\link{main__examples_analysis}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
