% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{main__examples_analysis}
\alias{main__examples_analysis}
\title{Pull all analysis types together}
\usage{
main__examples_analysis(what = c("aa", "sa", "uc", "ac"))
}
\arguments{
\item{what}{(character vector) what to process}
}
\value{
(data frame) attribution events
}
\description{
Pull all analysis types together
}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__aa_morbidity}},
  \code{\link{main__example_analysis__ac_morbidity}},
  \code{\link{main__example_analysis__sa_morbidity}},
  \code{\link{main__example_analysis__uc_morbidity}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
