% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct tables}
\usage{
reconstruct(this_version = c("aaf_2017_phe", "aaf_2014_ljucph",
  "aaf_2008_ljucph", "aaf_2007_ni39"), this_analysistype = c("morbidity",
  "mortality"))
}
\arguments{
\item{this_version}{table version to reconstruct}

\item{this_analysistype}{specify analysis type: either morbidity or mortality}
}
\description{
Reconstruct alcohol attributable fractions table for given version and analysis type.
}
