% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{ricd10}
\alias{ricd10}
\title{Return a random icd10 code.}
\usage{
ricd10(n = 1024, mix_type = c("5050aalen34mix", "len3", "len4",
  "len34", "aa_only"), nmultiple = 1)
}
\arguments{
\item{n}{(integer) number of records to return}

\item{mix_type}{(character) Specify code mix - len3:all Xnn, len4:all Xnnn,
len34:any Xnn/Xnnn, aa_only:only alcohol attributable related, or
5050aalen34mix a 50/50 mix of aa_only and len34}

\item{nmultiple}{(integer) return up to this many codes separated by ";" in
  each record.  Default 1 (and no ";").

Sampling from generated code space used, however for 50/50 mix due to problme
space the icd10 codesa re generated randomly on-the-fly rather than sampled
from full generated pattern space.  Still random but might not be the same
random ... but probably is.}
}
\value{
(character vector) n records of 1 to nmultiple (varying) icd10-like
  codes.
}
\description{
Return a random icd10 code.
}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__hes_ip}}
}
\concept{examples_of_analysis}
