% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_conditions}
\alias{sa_conditions}
\title{List of smoking attributable conditions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 6 columns.}
\usage{
sa_conditions
}
\description{
Provides lookup tables for use with smoking attributable fractions analyses.
}
\details{
# @format data frame with 26 rows and 4 fields

\preformatted{
Observations: 26
Variables: 4
$ cat1             <fct> Cancers which can be caused by smoking, Cancers which can be caused by smoking, Cancer...
$ disease_category <fct> "Trachea, Lung, Bronchus", "Upper respiratory sites", "Oesophagus", "Larynx", "Cervica...
$ icd_10_code      <fct> "C33-C34", "C00-C14", "C15", "C32", "C53", "C67", "C64-C66,C68", "C16", "C25", "C80", ...
$ condition_uid    <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24,...
}
}
\examples{
if (isNamespaceLoaded("magrittr")) {
  require("magrittr")

  aafractions.ncc::sa_conditions \%>\%
    head(16)
}

}
\seealso{
Other smoking datasets: \code{\link{lu_sac_icd10}},
  \code{\link{sa_relrisk}}, \code{\link{smoking_datasets}},
  \code{\link{sp}}
}
\concept{smoking datasets}
\keyword{datasets}
