% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_relrisk}
\alias{sa_relrisk}
\title{List of smoking relative risks by condition, age and sex}
\format{data frame with 1,016 rows and 8 fields

\preformatted{
Observations: 1,016
Variables: 8
$ age            <fct> 35 - 54, 35 - 54, 35 - 54, 35 - 54, 35 - 54, 35 - 54, 35 - 54, 35 - 54, 35 - 54, 35 - 54...
$ condition_uid  <int> 16, 16, 16, 16, 16, 16, 16, 16, 18, 18, 18, 18, 18, 18, 18, 18, 14, 14, 14, 14, 14, 14, ...
$ sex            <chr> "men", "men", "men", "men", "women", "women", "women", "women", "men", "men", "men", "me...
$ smoking_status <chr> "current", "current", "ex", "ex", "current", "current", "ex", "ex", "current", "current"...
$ ab_sa_explode  <chr> "35 - 44", "45 - 54", "35 - 44", "45 - 54", "35 - 44", "45 - 54", "35 - 44", "45 - 54", ...
$ analysis_type  <chr> "morbidity", "morbidity", "morbidity", "morbidity", "morbidity", "morbidity", "morbidity...
$ srr            <dbl> 4.20, 4.20, 2.00, 2.00, 5.30, 5.30, 2.60, 2.60, 4.40, 4.40, 1.10, 1.10, 5.40, 5.40, 1.30...
$ version        <chr> "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss...#'
}}
\usage{
sa_relrisk
}
\description{
Provides lookup tables for use with smoking attributable fractions analyses.
}
\examples{
if (isNamespaceLoaded("dplyr")) {
  require("dplyr")

  aafractions.ncc::sa_relrisk \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_")) \%>\%
    summary(16)
}

}
\seealso{
Other smoking datasets: \code{\link{lu_sac_icd10}},
  \code{\link{sa_conditions}},
  \code{\link{smoking_datasets}}, \code{\link{sp}}
}
\concept{smoking datasets}
\keyword{datasets}
