% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_versions}
\alias{sa_versions}
\title{List of smoking attributable conditions by version}
\format{data frame with 26 rows and 2 fields

\preformatted{
Observations: 26
Variables: 2
$ version       <chr> "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_2018", "nhsd_ss_...
$ condition_uid <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25...
}}
\usage{
sa_versions
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
if (isNamespaceLoaded("dplyr")) {
  require("dplyr")

  aafractions.ncc::sa_versions \%>\%
    dplyr::mutate_if(is.character, as.factor) \%>\%
    dplyr::select(-starts_with("condition_")) \%>\%
    summary(16)
}

}
\seealso{
Other alcohol datasets: \code{\link{aa_conditions}},
  \code{\link{aa_fractions}}, \code{\link{aa_versions}},
  \code{\link{alcohol_datasets}},
  \code{\link{lu_aac_icd10}}
}
\concept{alcohol datasets}
\keyword{datasets}
