% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smoking_datasets}
\alias{smoking_datasets}
\title{Smoking datasets}
\description{
Smoking datasets
}
\seealso{
Other smoking datasets: \code{\link{lu_sac_icd10}},
  \code{\link{sa_conditions}}, \code{\link{sa_relrisk}},
  \code{\link{sp}}
}
\concept{smoking datasets}
