% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp}
\alias{sp}
\title{Lookup for smoking prevalence}
\format{data frame with 19.012 rows and 8 fields}
\usage{
sp
}
\description{
To be combined with relative risk to arrive at attributable fraction
}
\examples{
head(sp)

}
\seealso{
Other smoking datasets: \code{\link{lu_sac_icd10}},
  \code{\link{sa_conditions}}, \code{\link{sa_relrisk}},
  \code{\link{smoking_datasets}}
}
\concept{smoking datasets}
\keyword{datasets}
