% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uc_attribution}
\alias{uc_attribution}
\title{List of Urgent care sensitive fractions by condition and age}
\format{data frame with 5,920 rows and 7 fields

\preformatted{
Observations: 38
Variables: 8
$ age                   <chr> "0 - 5 yrs", "75+ yrs", "All ages", "All ages", "All ages", "All ages", "All ages", "All ages",...
$ condition_description <chr> "Pyrexial child", "Falls", "COPD", "COPD", "COPD", "Acute mental health crisis", "Acute mental ...
$ primary_diagnosis     <chr> "R50", "W0; W1-W19", "J40; J41; J42; J43; J44", "J40; J41; J42; J43; J44", "J40; J41; J42; J43;...
$ primary_regex         <chr> "R50", "W[01][0-9]", "J4[0-4]", "J4[0-4]", "J4[0-4]", "F", "F", "F", "R07[2-4]", "R07[2-4]", "R...
$ condition_uid         <int> 8, 4, 1, 1, 1, 2, 2, 2, 3, 3, 3, 12, 12, 12, 5, 5, 5, 6, 6, 6, 7, 7, 7, 9, 9, 9, 10, 10, 10, 11...
$ version               <chr> "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_201617", "ccg_...
$ ab_ucs_explode        <chr> "0 - 5 yrs", "75+ yrs", "0 - 5 yrs", "6 - 74 yrs", "75+ yrs", "0 - 5 yrs", "6 - 74 yrs", "75+ y...
$ ucs_af                <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
}}
\usage{
uc_attribution
}
\description{
Provides lookup tables for use with Urgent care sensitive analyses.
}
\examples{
if (all(sapply(c("dplyr", "reshape2"), isNamespaceLoaded))) {
  require("dplyr", warn.conflicts = FALSE)
  require("reshape2", warn.conflicts = FALSE)

  uc_attribution \%>\%
    dcast(
      condition_uid + condition_description ~ ab_ucs_explode
      , value.var = "ucs_af", fill = 0
    )
}

}
\seealso{
Other urgent care sensitive datasets: \code{\link{lu_ucc_icd10}},
  \code{\link{uc_conditions}}, \code{\link{uc_versions}},
  \code{\link{ucs_datasets}}
}
\concept{urgent care sensitive datasets}
\keyword{datasets}
