% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uc_versions}
\alias{uc_versions}
\title{List of Urgent care sensitive conditions by version}
\format{data frame with 14 rows and 2 fields

\preformatted{
Observations: 14
Variables: 2
$ version       <chr> "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_201617", "ccg_iaf_2016...
$ condition_uid <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14
}}
\usage{
uc_versions
}
\description{
Provides lookup tables for use with Urgent care sensitive analyses.
}
\examples{
if (isNamespaceLoaded("dplyr")) {
  require("dplyr")

  aafractions.ncc::uc_versions \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_")) \%>\%
    summary(16)
}

}
\seealso{
Other urgent care sensitive datasets: \code{\link{lu_ucc_icd10}},
  \code{\link{uc_attribution}},
  \code{\link{uc_conditions}}, \code{\link{ucs_datasets}}
}
\concept{urgent care sensitive datasets}
\keyword{datasets}
