% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ucs_datasets}
\alias{ucs_datasets}
\title{Urgent care sensitive datasets}
\description{
Urgent care sensitive datasets
}
\seealso{
Other urgent care sensitive datasets: \code{\link{lu_ucc_icd10}},
  \code{\link{uc_attribution}},
  \code{\link{uc_conditions}}, \code{\link{uc_versions}}
}
\concept{urgent care sensitive datasets}
