% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_hesip.R
\name{ab_labels_from_breaks}
\alias{ab_labels_from_breaks}
\title{Create ageband labels}
\usage{
ab_labels_from_breaks(breaks)
}
\arguments{
\item{breaks}{(integer vector) breaks}
}
\value{
(character vector) labels
}
\description{
Given breaks construct ageband labels of the form "xx-yy Yrs" and "zz+
Yrs"
}
