% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{ab_labels_from_breaks}
\alias{ab_labels_from_breaks}
\title{Create ageband labels}
\usage{
ab_labels_from_breaks(breaks, style = c("alcohol", "smoking", "generic",
  "ucs"), flag = "0", width = 2)
}
\arguments{
\item{breaks}{(integer vector) breaks}

\item{style}{(character) choose style alcohol: "00-00 Yrs", smoking: "00 -
00", generic: "a0000"}

\item{flag}{(character) number formatting (see formatC)}

\item{width}{(integer) number formatting (see formatC)}
}
\value{
(character vector) labels
}
\description{
Given breaks construct ageband labels of the form "xx-yy Yrs" and "zz+
Yrs"
}
\details{
A general ageband has an age range, specfied by lower (incluseive) and upper
(exclusive).

An ageband label is of the general form:

<prefix> <value lower> <separator> <value upper> <suffix>

with the last age band of the form:

<prefix> <value upper> "+" <suffix>

Variations are tailored for specific lookup tables.

* alcohol: 00-00 Yrs
* smoking: 00 - 00
* generic: a0000
* ucs:     0 - 0 yrs
}
\seealso{
Other examples_of_analysis: \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__aa_morbidity}},
  \code{\link{main__example_analysis__sa_morbidity}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
