% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{create__dummy_hesip}
\alias{create__dummy_hesip}
\title{Create dummy HES table}
\usage{
create__dummy_hesip(n = 1024, mix_type = NULL)
}
\arguments{
\item{n}{(integer) length of table}

\item{mix_type}{(character) passed to ricd10 (random diagnosis code
generator)}
}
\value{
(data.frame) dummy HES table
}
\description{
Generate a psudo random HES IP table good enough to put any analysis through
its paces.
}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create_lu_ageband}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__aa_morbidity}},
  \code{\link{main__example_analysis__ac_morbidity}},
  \code{\link{main__example_analysis__sa_morbidity}},
  \code{\link{main__example_analysis__uc_morbidity}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
