% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_analysis__hes_ip.R
\name{create_lu_ageband}
\alias{create_lu_ageband}
\title{Create age band looks for aa and esp}
\usage{
create_lu_ageband(style = c("alcohol", "smoking", "generic"),
  name = NULL)
}
\arguments{
\item{style}{(character) choose label foramtting and breaks (alcohol,
smoking, generic - 5 year agebands to 95+)}

\item{name}{(character) choose field name}
}
\value{
(data.frame) with fields age, ab, ab_style ab optionally renamed)
}
\description{
Create age band looks for aa and esp
}
\examples{
require("dplyr")
aafractions.ncc:::create_lu_ageband() \%>\% mutate_if(is.character, as.factor) \%>\% summary(20)
aafractions.ncc:::create_lu_ageband() \%>\% select(starts_with("ab_")) \%>\% unique()

}
\seealso{
Other examples_of_analysis: \code{\link{ab_labels_from_breaks}},
  \code{\link{create__dummy_hesip}},
  \code{\link{create_lu_gender}},
  \code{\link{examples_of_analysis}},
  \code{\link{gen_codes}},
  \code{\link{main__example_analysis__hes_ip}},
  \code{\link{ricd10}}
}
\concept{examples_of_analysis}
