% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_aac_icd10}
\alias{lu_aac_icd10}
\title{Lookup between alcohol attributable conditions and icd10 codes}
\format{data frame with 69 rows and 7 fields}
\usage{
lu_aac_icd10
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\examples{
if (all(sapply(c("dplyr", "reshape2"), isNamespaceLoaded))) {
  require("dplyr")
  require("reshape2")

  aafractions.ncc::lu_aac_icd10 \%>\%
    merge(aafractions.ncc::aa_versions, by = "condition_uid", allow.cartesian = TRUE) \%>\%
    dcast(... ~ Version, value.var = "condition_uid", fun = paste, collapse = "|") \%>\%
    arrange(icd10) \%>\%
    head(16)
}

}
\seealso{
Other alcohol datasets: \code{\link{aa_conditions}},
  \code{\link{aa_fractions}}, \code{\link{aa_versions}},
  \code{\link{alcohol_datasets}}
}
\concept{alcohol datasets}
\keyword{datasets}
