% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_acc_opcs}
\alias{lu_acc_opcs}
\title{Lookup between Ambulatory care sensitive conditions and icd10 codes}
\format{data frame with 0 rows and 2 fields

\preformatted{
Observations: 0
Variables: 2
$ condition_uid <int>
$ icd10         <chr>
}}
\usage{
lu_acc_opcs
}
\description{
Provides lookup tables for use with Ambulatory care sensitive analyses.
Tertiary condition: procedure codes.
}
\examples{
str(aafractions.ncc::lu_acc_opcs)

}
\seealso{
Other ambulatory care sensitive datasets: \code{\link{ac_attribution}},
  \code{\link{ac_conditions}}, \code{\link{ac_versions}},
  \code{\link{acs_datasets}},
  \code{\link{lu_acc_icd10_sec}},
  \code{\link{lu_acc_icd10}}
}
\concept{ambulatory care sensitive datasets}
\keyword{datasets}
