% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_sac_icd10}
\alias{lu_sac_icd10}
\title{Lookup between Smoking attributable conditions and icd10 codes}
\format{data frame with 69 rows and 7 fields}
\usage{
lu_sac_icd10
}
\description{
Provides lookup tables for use with smoking relative risks / attributable
fractions analyses.
}
\examples{
if (all(sapply(c("dplyr", "reshape2"), isNamespaceLoaded))) {
  require("dplyr")
  require("reshape2")

  aafractions.ncc::lu_sac_icd10 \%>\%
    merge(aafractions.ncc::sa_conditions, by = "condition_uid", allow.cartesian = TRUE) \%>\%
    dcast(... ~ version, value.var = "condition_uid", fun = paste, collapse = "|") \%>\%
    arrange(icd10) \%>\%
    head(16)
}

}
\seealso{
Other smoking datasets: \code{\link{sa_conditions}},
  \code{\link{sa_relrisk}}, \code{\link{smoking_datasets}},
  \code{\link{sp}}
}
\concept{smoking datasets}
\keyword{datasets}
