% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_versions}
\alias{lu_versions}
\title{List of alcohol attributable fractions by condition, age and sex}
\format{data frame with 221 rows and 10 fields

'data.frame':	221 obs. of  10 variables:
$ Version      : Factor w/ 4 levels "aaf_2007_ni39",..: 4 4 4 4 4 4 4 4 4 4 ...
$ cat1         : Factor w/ 3 levels "Partially attributable conditions - acute conditions",..: 3 3 3 3 3 3 3 3 3 3 ...
$ cat2         : Factor w/ 12 levels "Cardiovascular disease",..: 12 12 12 12 12 12 12 12 12 12 ...
$ desc         : Factor w/ 68 levels "§","§§","Accidental excessive cold",..: 9 49 21 14 13 10 11 12 7 8 ...
$ attribution  : logi  NA NA NA NA NA NA ...
$ cause        : logi  NA NA NA NA NA NA ...
$ analysis_type: Factor w/ 3 levels "all","morbidity",..: 1 1 1 1 1 1 1 1 1 1 ...
$ sortkey      : num  1 1 1 1 1 1 1 1 1 1 ...
$ codes        : Factor w/ 71 levels "§","§§","§§ Rest of 'V' series",..: 14 15 16 18 19 22 30 31 37 38 ...
$ gucid        : num  1 2 3 4 5 6 7 8 9 10 ...}
\usage{
lu_versions
}
\description{
Provides lookup tables for use with alcohol attributable fractions analyses.
}
\keyword{datasets}
