% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct lu tables}
\usage{
reconstruct(this_version = c("aaf_2017_phe", "aaf_2014_ljucph",
  "aaf_2008_ljucph", "aaf_2007_ni39"), this_analysistype = c("morbidity",
  "mortality"), verbose = TRUE, molten = FALSE)
}
\arguments{
\item{this_version}{table version to reconstruct}

\item{this_analysistype}{specify analysis type: either morbidity or mortality}

\item{verbose}{show messages about which table is being reconstructed}

\item{molten}{boolean to indicate cast or molten on (sex, ageband)}
}
\description{
Reconstruct alcohol attributable fractions table for given version and
analysis type.
}
\examples{

t1 <- reconstruct("aaf_2017_phe", "morbidity")
str(t1)

t2 <- reconstruct("aaf_2017_phe", "morbidity", molten = TRUE)
str(t2)


}
