% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_relrisk}
\alias{sa_relrisk}
\title{List of smoking relative risks by condition, age and sex}
\format{data frame with 140 rows and 5 fields

\preformatted{
Observations: 140
Variables: 5
$ condition_uid  <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 14, 15, 16, 16, 16, 16, 17, 18, 18, 18, 1...
$ age            <fct> 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35+, 35 - 64, 65+, 35+, 35 -...
$ sex            <chr> "men", "men", "men", "men", "men", "men", "men", "men", "men", "men", "men", "men", "men...
$ smoking_status <chr> "current_smokers", "current_smokers", "current_smokers", "current_smokers", "current_smo...
$ srr            <dbl> 23.26, 10.89, 6.76, 14.60, 1.00, 3.27, 2.50, 1.96, 2.31, 4.40, 1.80, 17.10, 10.58, 2.50,...
}}
\usage{
sa_relrisk
}
\description{
Provides lookup tables for use with smoking attributable fractions analyses.
}
\examples{
if (isNamespaceLoaded("dplyr")) {
  require("dplyr")

  aafractions.ncc::sa_relrisk \%>\%
    mutate_if(is.character, as.factor) \%>\%
    select(-starts_with("condition_")) \%>\%
    summary(16)
}

}
\seealso{
Other smoking datasets: \code{\link{lu_sac_icd10}},
  \code{\link{sa_conditions}},
  \code{\link{smoking_datasets}}, \code{\link{sp}}
}
\concept{smoking datasets}
\keyword{datasets}
