% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp}
\alias{sp}
\title{Lookup for smoking prevalence}
\format{data frame with 19,012 rows and 9 fields

\preformatted{
Observations: 19,012
Variables: 9
$ smoking_status <chr> "current", "current", "current", "current", "current", "current", "current", "current", ...
$ calyear        <int> 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011, 2011...
$ area_code      <chr> "E92000001", "E92000001", "E92000001", "E92000001", "E92000001", "E92000001", "E92000001...
$ area_name      <chr> "England", "England", "England", "England", "England", "England", "England", "England", ...
$ area_type      <chr> "England", "England", "England", "England", "England", "England", "England", "England", ...
$ sex            <chr> "M", "F", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P", "P"...
$ age            <chr> "18+ yrs", "18+ yrs", "25-29 yrs", "30-34 yrs", "35-39 yrs", "40-44 yrs", "45-49 yrs", "...
$ value          <dbl> 22.1785, 17.6335, 27.1005, 24.1422, 23.3574, 22.5773, 21.5718, 20.1515, 19.4923, 16.2628...
$ version        <chr> "phe_ltcp_201903", "phe_ltcp_201903", "phe_ltcp_201903", "phe_ltcp_201903", "phe_ltcp_20...
}}
\usage{
sp
}
\description{
To be combined with relative risk to arrive at attributable fraction
}
\examples{
head(sp)

}
\seealso{
Other smoking datasets: \code{\link{lu_sac_icd10}},
  \code{\link{sa_conditions}}, \code{\link{sa_relrisk}},
  \code{\link{smoking_datasets}}
}
\concept{smoking datasets}
\keyword{datasets}
